Data Validation

+++ api/Controllers/CommentController.cs
@@ -27,15 +27,19 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetComments()
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var commentsModel = await _commentRepo.GetCommentsAsync();
             var CommentsDto = commentsModel.Select(s => s.ToCommentDto());
 
             return Ok(commentsModel);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetComment([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var commentModel = await _commentRepo.GetCommentAsync(id);
 
             if (commentModel == null) return NotFound();
@@ -43,9 +47,11 @@ namespace api.Controllers
             return Ok(commentModel.ToCommentDto());
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> CreateComment([FromRoute] int stockId, CreateCommentRequestDto commentDto)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             if (!await _stockRepo.StockExists(stockId))
                 return BadRequest("Stock does not exist");
 
@@ -56,9 +62,11 @@ namespace api.Controllers
             return CreatedAtAction(nameof(GetComment), new { id = commentModel.Id }, commentModel.ToCommentDto());
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("{id:int}")]
         public async Task<IActionResult> UpdateComment([FromRoute] int id, [FromBody] UpdateCommentRequestDto commentDto)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             Comment? commentModel = await _commentRepo.UpdateCommentAsync(id, commentDto);
 
             if (commentModel == null) return NotFound();
@@ -66,9 +74,11 @@ namespace api.Controllers
             return Ok(commentModel.ToCommentDto());
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> DeleteComment([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             Comment? commentModel = await _commentRepo.DeleteCommentAsync(id);
 
             if (commentModel == null) return NotFound();
